requirements.txt

PyPDF2==3.0.1
python-dotenv==1.0.0
streamlit==1.18.1
faiss-cpu==1.7.4
altair==4
tiktoken==0.4.0
# uncomment to use huggingface llms
huggingface-hub==0.14.1

# uncomment to use instructor embeddings
InstructorEmbedding==1.0.1
langchain==0.1.17
langchain-community==0.0.38
langchain-text-splitters==0.0.1
langsmith==0.1.52
sentence-transformers==2.7.0
tokenizers==0.19.1
transformers==4.40.2